requirements.txt
alabaster==0.7.12
alembic==1.7.6
altair==4.2.0
anyio==3.5.0
apache-airflow==2.2.4
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.3
apache-airflow-providers-imap==2.2.0
apache-airflow-providers-sqlite==2.1.0
apispec==3.3.2
appdirs==1.4.4
appnope==0.1.2
argcomplete==1.12.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
asttokens==2.0.5
attrs==20.3.0
autopep8==1.6.0
aws==0.2.5
awscli==1.22.89
Babel==2.9.1
backcall==0.2.0
base58==2.1.1
bcrypt==3.2.0
bleach==4.1.0
blinker==1.4
boto==2.49.0
boto3==1.21.34
botocore==1.24.34
Brotli==1.0.9
cachelib==0.6.0
cachetools==5.0.0
cattrs==1.10.0
certifi==2020.12.5
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
commonmark==0.9.1
ConfigArgParse==1.5.3
connexion==2.11.1
croniter==1.3.4
cryptography==3.4.8
cycler==0.11.0
Cython==0.29.28
debugpy==1.5.1
decorator==4.3.2
defusedxml==0.7.1
Deprecated==1.2.13
descartes==1.1.0
dill==0.3.1.1
dnspython==2.2.0
docutils==0.16
elastic-transport==8.1.0
elasticsearch==7.16.2
elasticsearch-dsl==7.4.0
elyzer==1.2.2
email-validator==1.1.3
entrypoints==0.3
esbonio==0.11.0
executing==0.8.3
fabric==2.6.0
Flask==1.1.2
Flask-AppBuilder==3.4.4
Flask-Babel==2.0.0
Flask-BasicAuth==0.2.0
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.3.0
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
fonttools==4.29.1
fuzzywuzzy==0.18.0
gevent==21.12.0
geventhttpclient==1.5.3
gitdb==4.0.9
GitPython==3.1.27
graphviz==0.19.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
imagesize==1.3.0
importlib-metadata==4.11.1
inflection==0.5.1
invoke==1.6.0
ipykernel==6.4.2
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
iso8601==1.0.2
isodate==0.6.1
itsdangerous==1.1.0
jarowinkler==1.0.2
jedi==0.18.1
Jinja2==3.0.3
jmespath==1.0.0
joblib==1.1.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.6
Markdown==3.3.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
mizani==0.7.3
msgpack==1.0.3
nbclient==0.5.12
nbconvert==5.6.1
nbformat==5.1.3
nbgrader==0.6.2
nbstripout==0.5.0
nest-asyncio==1.5.4
notebook==6.4.8
numpy==1.21.5
numpydoc==1.2.1
openapi-schema-validator==0.1.6
openapi-spec-validator==0.3.3
packaging==21.3
palettable==3.3.0
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.9.2
parso==0.8.3
pathlib2==2.3.7.post1
patsy==0.5.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plotly==5.5.0
plotnine==0.8.0
prettytable==3.2.0
prison==0.2.1
prometheus-client==0.13.1
prompt-toolkit==3.0.26
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==7.0.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.8.2
pydeck==0.7.1
pygls==0.11.3
Pygments==2.11.2
PyJWT==1.7.1
Pympler==1.0.1
PyNaCl==1.5.0
pyparsing==2.4.7
pyrsistent==0.16.1
pyspellchecker==0.6.3
python-daemon==2.3.0
python-dateutil==2.8.2
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pytzdata==2020.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
rapidfuzz==2.0.10
requests==2.27.1
retrying==1.3.3
rfc3986==1.5.0
rich==11.2.0
roundrobin==0.0.2
rsa==4.7.2
s3transfer==0.5.2
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.11.2
semver==2.13.0
Send2Trash==1.8.0
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
Sphinx==4.5.0
sphinx-rtd-theme==1.0.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.38.2
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.7.0
swagger-ui-bundle==0.0.9
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==4.3.3
typeguard==2.13.3
typing_extensions==4.2.0
tzdata==2021.5
tzlocal==4.1
unicodecsv==0.14.1
urllib3==1.26.8
validators==0.18.2
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.2
wrapt==1.13.3
WTForms==2.3.3
xgboost==1.5.2
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
